rough baseline estimation, rdf::Line cleaned (float -> double)

+++ src/Core/Blobs.h
@@ -132,7 +132,7 @@ public:
 	static BlobManager& instance();
 	QVector<Blob> filterArea(int area, const Blobs& blobs) const;
 	QVector<Blob> filterMar(float maxAspectRatio, int minWidth, const Blobs& blobs) const;
-	QVector<Blob> filterAngle(float angle, float maxAngleDiff, const Blobs& blobs) const;
+	QVector<Blob> filterAngle(double angle, double maxAngleDiff, const Blobs& blobs) const;
 	cv::Mat drawBlobs(const Blobs& blobs, cv::Scalar color = cv::Scalar(255, 255, 255)) const;
 	QVector<Line> lines(const Blobs& blobs) const;
 	Blob getBiggestBlob(const Blobs& blobs) const;
